Moved everything into new structure, reworked API, moving to laptop now
